# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk aplikasi internal",
        "title": "Aplikasi Internal API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/permintaan": {
            "get": {
                "description": "Mendapatkan daftar permintaan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permintaan"
                ],
                "summary": "Ambil semua permintaan",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/permintaan.Permintaan"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Menambahkan data permintaan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permintaan"
                ],
                "summary": "Membuat permintaan baru",
                "parameters": [
                    {
                        "description": "Data permintaan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/permintaan.PermintaanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/permintaan.Permintaan"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/permintaan-nama": {
            "get": {
                "description": "Mendapatkan daftar permintaan yang sudah tertampil nama",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permintaan"
                ],
                "summary": "Ambil semua permintaan yang sudah tertampil nama",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/permintaan.PermintaanByNama"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
        "/permintaan-nama/{id}": {
            "get": {
                "description": "Mendapatkan data permintaan yang sudah tertampil nama berdasarkan UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permintaan"
                ],
                "summary": "Ambil permintaan yang sudah tertampil nama berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Permitaan ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/permintaan.PermintaanByNama"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/permintaan/{id}": {
            "get": {
                "description": "Mendapatkan data permintaan berdasarkan UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permintaan"
                ],
                "summary": "Ambil permintaan berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Permitaan ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/permintaan.Permintaan"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Mengupdate data permintaan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permintaan"
                ],
                "summary": "Update permintaan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Permintaan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data permintaan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/permintaan.PermintaanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/permintaan.Permintaan"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus data permintaan berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permintaan"
                ],
                "summary": "Hapus permintaan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Permintaan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/helpers.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "helpers.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "permintaan.Permintaan": {
            "type": "object",
            "properties": {
                "aplikasi_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kondisi_awal": {
                    "type": "string"
                },
                "kondisi_diharapkan": {
                    "type": "string"
                },
                "menu": {
                    "type": "string"
                },
                "pemda_id": {
                    "type": "string"
                },
                "tanggal_deadline": {
                    "type": "string"
                },
                "tanggal_pesanan": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "permintaan.PermintaanByNama": {
            "type": "object",
            "properties": {
                "aplikasi": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kondisi_awal": {
                    "type": "string"
                },
                "kondisi_diharapkan": {
                    "type": "string"
                },
                "menu": {
                    "type": "string"
                },
                "pembuat": {
                    "type": "string"
                },
                "pemda": {
                    "type": "string"
                },
                "tanggal_deadline": {
                    "type": "string"
                },
                "tanggal_pesanan": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "permintaan.PermintaanRequest": {
            "type": "object",
            "required": [
                "aplikasi_id",
                "created_by",
                "kondisi_awal",
                "kondisi_diharapkan",
                "menu",
                "pemda_id",
                "tanggal_deadline",
                "tanggal_pesanan"
            ],
            "properties": {
                "aplikasi_id": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "kondisi_awal": {
                    "type": "string",
                    "minLength": 3
                },
                "kondisi_diharapkan": {
                    "type": "string",
                    "minLength": 3
                },
                "menu": {
                    "type": "string",
                    "minLength": 3
                },
                "pemda_id": {
                    "type": "string"
                },
                "tanggal_deadline": {
                    "type": "string"
                },
                "tanggal_pesanan": {
                    "type": "string"
                }
            }
        },
...
    }
}